pyproject.toml
[project]
name = "czech-banking-ci"
version = "0.2.0"
description = "Czech Banking Competitive Intelligence Monitor"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "tenacity>=8.2",
    "python-dateutil>=2.9",
    "openpyxl>=3.1",
    "feedparser>=6.0",
    "flask>=3.0",
    "flask-cors>=5.0",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
ci-monitor = "src.cli:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 110
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "UP",       # pyupgrade
    "B",        # flake8-bugbear
    "SIM",      # flake8-simplify
]
ignore = [
    "E501",     # line length; soft enforced by formatter
    "B008",     # function calls in defaults (fine for Click)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011", "F401"]
